Merge branch 'diana' into develop
Added the sidebar for the admin view, which is responsive. Also, I made
the nabber for the user responsive. The only thing I need to fix for
that is the search bar. Last, but not least, I changed the button that
is on the resources in the user view so it would be more visible.