User booking 95%
The user can now book a resource, and depending on its status it can
either book it or enter to the waiting list.

It’s almost finished, i need to upgrade the controller in order to fix
the concurrency problem (if the user saves the booking and the resource
gets free, then the controller doesn’t detect it and it save it as
waiting list, need to fix that).