90%: Search resource requirement
Add: Now the user can search resources using the search box in the
menu. Some structure of the project was changed in order to fix the
slide bug. Some routes were also changed in order to give it some good
structure. There is a bug: in the resource show view, when trying to
change lab, it doesn’t works. It’s because of the divs. Since the
jquery script needs the #resource class to fill the data it doesn’t
works in booking view because there is no resource div. I’m gonna fix
that at the end of the week.