Reassign resource requirement done
Now the admin can:
-Visualize a list of the active bookings (i considered that it wasn’t
important to show the past bookings so I just show the ones that are
active)
-Reassign the resource by clicking over the corresponding button.
-The system sends 2 emails, one for the first user to notify about the
cancelation and one to the first user in the waiting list to notify
about the new assignation.
-The system reassign the resource to the corresponding user
-The admin can visualize a list of the waiting list resources/users and
can erase them ( I considered this was possible)
Note: both waiting list and bookings are not complete CRUDS, since you
cannot create bookings or edit them (just reassign them).

EXTRA: I fixed the problem with the admin left bar, now the whole
“label” is clickeable. I just changed the css so the li display
attribute were set to block.